requirements.txt
asgiref @ file:///C:/ci_311/asgiref_1676448296195/work
Django @ file:///C:/ci_311/django_1676469589007/work
django-filter==23.3
djangorestframework==3.14.0
Pillow==10.1.0
pytz==2023.3.post1
sqlparse @ file:///C:/b/abs_88361ub_qu/croot/sqlparse_1690904577514/work
tzdata @ file:///croot/python-tzdata_1690578112552/work
